Update check on debug config generation fail
When sketch debugging is initiated, the extension runs Arduino CLI to get the debugger configuration information. This
information is required to generate the `launch.json` configuration file.

As part of this information is the path to the compiled sketch binary, Arduino CLI checks for the existence of that
file, and errors if it is not present (which is usually caused by the user having neglected to compile the sketch):

```
Error getting Debug info: Compiled sketch not found in ...
```

The extension has special handling for this specific common error cause, which is based on a fragile string comparison
on the human readable error message. As is the nature of this approach, a seemingly innocuous capitalization change to
the message in the Arduino CLI code base broke the check.

The quick fix provided here is making the check case insensitive.