Add test coverage for overriding debug configuration via debug_custom.json
Some of the data from the `arduino-cli debug --info` output does not match exactly with the launch.json format required
by the IDE's debugger. For this reason, some remapping of the data is required.

The user can adjust the debugger configuration via a debug_custom.json file in the sketch project. The data from this
file is merged into the base data provided by `arduino-cli debug --info`. It is essential that the data from
debug_custom.json overrides the base data where there is overlap, yet previously there was no test coverage for this.

Since the merging code gives special handling to the remapped Arduino CLI data, it is especially important to provide
good coverage for overriding that data.