Allow overriding remapped debug configuration via debug_custom.json
Some of the data from the `arduino-cli debug --info` output does not match exactly with the launch.json format required
by the IDE's debugger. For this reason, some remapping of the data is required.

The user can adjust the debugger configuration via a debug_custom.json file in the sketch project. The data from this
file is merged into the base data provided by `arduino-cli debug --info`. The data from debug_custom.json should
override the base data where there is overlap.

Previously the override did not work for the remapped Arduino CLI data since the remapping was applied after merging the
debug_custom.json data, which caused the debug_custom.json data to be overridden by that subset of the Arduino CLI data.
The bug is fixed by performing the remapping before merging the debug_custom.json data.