Fix bug in Arduino CLI debug config remapping tests
Some of the data from the `arduino-cli debug --info` output does not match exactly with the launch.json format required
by the IDE's debugger. For this reason, some remapping of the data is required. The remapping code is validated by
tests.

Previously the tests passed the remapping test data via the argument used for the data from the user's debug_custom.json
file. This was incorrect since the remapping is specific to the data from Arduino CLI. No remapping is necessary for the
debug_custom.json data since this is directly correlated to launch.json and thus will always use the correct data format
for direct inclusion in launch.json.